requirements.txt

# Python 3.13.7+

# PDF  
pypdf==4.3.1
reportlab==4.2.0

# Imagenes  
Pillow==10.4.0
qrcode==7.4.2

# Excel  
pandas==2.2.2
openpyxl==3.1.5

# HTTP para las peticiones 
requests==2.32.3

numpy==1.26.4  # para pandas
python-dateutil==2.9.0   
